refactor(ui): split AppUi
