+++ src/ui/mod.rs
@@ -1,35 +1,37 @@
 //! The egui layer: [`AppUi`] owns the gamepad cursor and the overlay state
-//! holders ([`crate::overlay::menu`], [`crate::overlay::osk`]), and lays the
-//! chrome out over the page. egui itself lives in
+//! holders, and lays the chrome out over the page. egui itself lives in
 //! [`crate::platform::window::AppWindow`], which owns the renderer it belongs
-//! to. The actual widgets are rendered by the submodules: [`toolbar`], [`menu`]
-//! (the full-screen overlay), and [`osk`].
+//! to; the widgets are rendered by the submodules ([`toolbar`], [`menu`], [`osk`]).
 
+mod cursor;
 mod dial_edit;
 mod hints;
 mod home;
 mod memory;
 mod menu;
 mod osk;
+mod overlays;
 mod panel;
 mod prompt;
+mod scale;
 mod settings;
 mod theme;
 mod toolbar;
 
+pub use self::overlays::Focus;
+
 use crate::{
-    app::{AppCommand, SettingsAction},
+    app::AppCommand,
     browser::AppBrowser,
     config::{
-        AppConfig, DebugConfig, DisplayConfig, DownloadsConfig, HistoryConfig, InputConfig,
-        OskConfig, ToolbarPosition, UpdateConfig,
+        DebugConfig, DisplayConfig, DownloadsConfig, HistoryConfig, InputConfig, OskConfig,
+        ToolbarPosition, UpdateConfig,
     },
-    event::user::UserEventSender,
-    overlay::dial_edit::{DialEdit, EditItem},
-    overlay::hints::{Hint, HintInput, HintLabels, Hints, Label, Sym},
+    overlay::dial_edit::DialEdit,
+    overlay::hints::Hints,
     overlay::home::Home,
     overlay::menu::Menu,
-    overlay::osk::{Osk, OskCommand, OskTarget},
+    overlay::osk::Osk,
     overlay::prompt::Prompt,
     overlay::settings::Settings,
     platform::window::AppWindow,
@@ -38,23 +40,9 @@ use crate::{
 use egui_sdl2::egui;
 use std::time::{Duration, Instant};
 
-/// The panel the chrome was drawn for; every other one is that, scaled. A pair
-/// of edges, not a width and a height: [`wanted_scale`] fits it either way round.
-const BASE_SIZE: egui::Vec2 = egui::vec2(640.0, 480.0);
-/// Bounds on the zoom actually installed, whatever the fit and the setting
-/// between them ask for.
-const SCALE_RANGE: std::ops::RangeInclusive<f32> = 0.5..=4.0;
-const SCALE_STEP: f32 = crate::config::bounds::SCALE_STEP as f32;
-/// A panel reported past this is egui's placeholder rect, not a screen; the real
-/// size arrives with the first frame.
-const MAX_PANEL: f32 = 8192.0;
-/// How far past a whole number a fit may reach before the fraction is worth it:
-/// fractional zoom lands glyphs between pixels, and the spare pixels become page.
-const WHOLE_ZOOM_REACH: f32 = 0.25;
-
 /// Style a freshly built [`egui::Context`]: the accent theme plus a launcher's
 /// pinned zoom. The fit to the panel needs a frame's measurements and arrives
-/// with [`AppUi::sync_scale`].
+/// with `sync_scale` (see [`scale`]).
 pub fn init_egui_ctx(ctx: &egui::Context) {
     theme::apply(ctx);
     if let Some(scale) = crate::config::device_scale() {
@@ -63,11 +51,9 @@ pub fn init_egui_ctx(ctx: &egui::Context) {
 }
 
 /// The text field the OSK currently types into, so its renderer can park egui's
-/// caret at the buffer end. The OSK only appends / backspaces, but egui keeps its
-/// own caret keyed by widget id and won't follow an external edit — left alone it
-/// sticks at the start and typed text scrolls out of view. `None` when the OSK is
-/// hidden or types somewhere without an egui caret (the page, or a settings row,
-/// whose value is painted text not a `TextEdit`).
+/// caret at the buffer end — egui keys its caret by widget id and won't follow an
+/// external edit. `None` when the OSK is hidden or types somewhere without an
+/// egui caret (the page, or a settings row, whose value is painted text).
 #[derive(PartialEq, Eq, Clone, Copy)]
 pub(super) enum OskField {
     None,
@@ -77,10 +63,6 @@ pub(super) enum OskField {
     Home,
 }
 
-/// Park egui's caret at char index `pos` (clamped to `char_count`) in an
-/// externally-edited single-line `TextEdit`, so it tracks the OSK's caret (see
-/// [`OskField`]). Call it *before* the field renders so the caret lands this
-/// frame; egui reloads the state we store here when it draws.
 /// Drop egui's own keyboard focus, so Enter reaches only the overlay's selected
 /// row (see the callers).
 fn drop_egui_focus(ctx: &egui::Context) {
@@ -91,6 +73,9 @@ fn drop_egui_focus(ctx: &egui::Context) {
     });
 }
 
+/// Park egui's caret at char index `pos` (clamped to `char_count`) in an
+/// externally-edited single-line `TextEdit`, so it tracks the OSK's caret (see
+/// [`OskField`]). Call *before* the field renders so the caret lands this frame.
 pub(super) fn park_caret(ctx: &egui::Context, id: egui::Id, pos: usize, char_count: usize) {
     let mut state = egui::TextEdit::load_state(ctx, id).unwrap_or_default();
     let at = egui::text::CCursor::new(pos.min(char_count));
@@ -110,41 +95,6 @@ pub(super) fn select_all(ctx: &egui::Context, id: egui::Id, char_count: usize) {
     );
     state.cursor.set_char_range(Some(range));
     egui::TextEdit::store_state(ctx, id, state);
-}
-
-/// Gamepad cursor overlay: circle radius and outline width (logical px).
-const CURSOR_RADIUS: f32 = 5.0;
-const CURSOR_STROKE: f32 = 1.5;
-/// The cursor's full painted half-extent — how far it reaches from its center,
-/// used to keep the whole glyph (not just the center) inside the web view.
-const CURSOR_EXTENT: f32 = CURSOR_RADIUS + CURSOR_STROKE / 2.0;
-
-/// Which surface owns contextual input (Confirm, Cancel, overlay `Nav` steps)
-/// right now: one precedence order derived from the overlay visibility flags,
-/// so the router and event handlers match on this instead of re-combining
-/// `*_visible()` checks at every site. The on-screen keyboard outranks the
-/// modal prompt (it's how a gamepad types into one), the prompt outranks the
-/// user overlays; menu / keyboard / hints never coexist (opening one closes
-/// the others).
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub enum Focus {
-    /// The on-screen keyboard — above everything, including the modal prompt.
-    Osk,
-    /// A modal page prompt (select picker / JS dialog) with no keyboard over it.
-    Prompt,
-    /// The full-screen menu (Tabs / Bookmarks / History / Downloads).
-    Menu,
-    /// The full-screen settings overlay (the on-screen keyboard can open over it
-    /// to type into a text field, hence it ranks below `Osk`).
-    Settings,
-    /// Link-hint navigation.
-    Hints,
-    /// The standalone speed-dial editor (opened from the start page).
-    DialEdit,
-    /// The built-in start page overlay (active tab is on `retsurf:home`).
-    Home,
-    /// No overlay: input goes to the page or the toolbar.
-    Page,
 }
 
 /// Toolbar layout decided before the egui closure ([`AppUi::toolbar_layout`]):
@@ -174,19 +124,17 @@ struct FrameInputs {
 }
 
 pub struct AppUi {
-    /// The window's egui context, cached so the focus/scale queries scattered
-    /// across the UI and the event handler don't each need the window. Refreshed
-    /// every frame: the software backend builds a fresh context on a resize.
+    /// The window's egui context, cached so focus/scale queries don't each need
+    /// the window. Refreshed every frame: the software backend builds a fresh
+    /// context on a resize.
     egui_ctx: egui::Context,
     repaint_delay: Option<Duration>,
     /// The web view's rect (logical px), measured from the central panel each
-    /// frame. With a top toolbar it sits below the bar; with a bottom toolbar it
-    /// starts at the window top. Used to map cursor/browser coordinates, keep
-    /// the cursor/pointer out of the toolbar, and anchor the home/hints overlays.
+    /// frame. Maps cursor/browser coordinates, keeps the pointer out of the
+    /// toolbar, and anchors the home/hints overlays.
     webview_rect: egui::Rect,
     /// Toolbar thickness (logical px), measured each frame. Stays valid across
-    /// window-size changes (unlike the rect, whose extent moves with the window),
-    /// so the SDL-resize fast path uses it to size the viewport without lag.
+    /// window-size changes, so the SDL-resize fast path sizes the viewport with it.
     toolbar_height: f32,
     /// The toolbar's current on-screen rect (logical px) — the panel strip, or
     /// the auto-hide overlay's slid position (off-screen when hidden). Used by the
@@ -207,7 +155,7 @@ pub struct AppUi {
     /// Last browser viewport size (physical px) we requested, to avoid churn.
     browser_viewport: (u32, u32),
     /// Gamepad cursor position (logical px). The UI owns it — it draws the
-    /// overlay — and the gamepad moves it via [`AppUi::move_cursor`].
+    /// overlay — and the gamepad moves it via `move_cursor` (see [`cursor`]).
     cursor: (f32, f32),
     /// When the cursor last moved, or `None` if it has never moved. Drives the
     /// auto-hide: the overlay shows only within `cursor_linger` of this.
@@ -223,9 +171,8 @@ pub struct AppUi {
     toolbar_position: ToolbarPosition,
     /// Whether the toolbar hides on scroll-down / reveals on scroll-up (config).
     toolbar_autohide: bool,
-    /// Auto-hide target visibility: shown after a scroll up, hidden after a scroll
-    /// down. Ignored unless `toolbar_autohide`; forced shown while a field is being
-    /// typed into (so the address bar is reachable).
+    /// Auto-hide target visibility: shown after a scroll up, hidden after a
+    /// scroll down; forced shown while a field is typed into.
     toolbar_shown: bool,
     /// Signed scroll distance accumulated since the last direction flip; the
     /// toolbar flips visibility once it crosses a threshold (debounces jitter).
@@ -238,8 +185,8 @@ pub struct AppUi {
     /// The full-screen settings overlay (edits a config draft). Public — driven
     /// directly; open/close/move go through the `settings_*` coordinators.
     pub settings: Settings,
-    /// Self-update manager (About tab): in-place on PortMaster / desktop installs,
-    /// "open the release page" elsewhere. See [`crate::update`].
+    /// Self-update manager (About tab): in-place on PortMaster / desktop
+    /// installs, "open the release page" elsewhere. See [`crate::update`].
     update: Updater,
     /// The built-in start page overlay's selection / search-field state.
     home: Home,
@@ -249,8 +196,7 @@ pub struct AppUi {
     /// [`crate::browser::AppBrowser::on_home_page`]); drives [`Focus::Home`].
     home_active: bool,
     /// Link-hint navigation state (L3); rects come from the browser. Public —
-    /// driven directly; a round starts via [`AppUi::hints_begin_collect`] and
-    /// [`AppUi::hints_apply`].
+    /// a round starts via `hints_begin_collect` / `hints_apply` (see [`overlays`]).
     pub hints: Hints,
     /// Modal page prompts: queued `<select>` pickers and JS dialogs. Public —
     /// the router and main loop drive [`Prompt`]'s own methods directly.
@@ -261,14 +207,13 @@ pub struct AppUi {
     /// Whether hint mode draws combo badges (cached from the input config; the
     /// router gates symbol routing on the same flag). Off = plain spatial hops.
     hint_badges: bool,
-    /// Whether the most recent input came from the keyboard (vs a gamepad), noted
-    /// by the event handler. Read when hint mode opens to pick the badge alphabet:
-    /// typed letters for the keyboard, button combos for the pad.
+    /// Whether the most recent input came from the keyboard (vs a gamepad).
+    /// Read when hint mode opens to pick the badge alphabet.
     last_input_keyboard: bool,
     /// Whether the debug memory overlay is enabled (`[debug] memory_overlay`).
     memory_overlay: bool,
-    /// Whether the same report goes to the log (`[debug] memory_log`), which is
-    /// how a handheld answers the question without covering its own screen.
+    /// Whether the same report goes to the log (`[debug] memory_log`) — how a
+    /// handheld answers the question without covering its own screen.
     memory_log: bool,
     /// The latest rolled-up memory report to draw, refreshed from the main loop.
     memory_summary: Option<memory::MemorySummary>,
@@ -356,9 +301,8 @@ impl AppUi {
         self.memory_summary = Some(memory::MemorySummary::from_report(report));
     }
 
-    /// Write the latest figures to the log, if `[debug] memory_log` asked for it
-    /// (see [`memory::MemorySummary::log`]). `compose_bytes` is the compositor's,
-    /// from [`crate::platform::AppWindow::compose_bytes`].
+    /// Write the latest figures to the log, if `[debug] memory_log` asked for it.
+    /// `compose_bytes` is the compositor's, from [`AppWindow::compose_bytes`].
     pub fn log_memory_summary(&self, ctx: &egui::Context, compose_bytes: usize) {
         if let (true, Some(summary)) = (self.memory_log, &self.memory_summary) {
             summary.log();
@@ -378,30 +322,20 @@ impl AppUi {
         self.repaint_delay.take()
     }
 
-    /// Ask the main loop to render one more frame immediately, without blocking on
-    /// input. Commands are drained *after* this frame's [`AppUi::update`] builds the
-    /// egui output, so a command that changes UI state (open a menu, switch section,
-    /// type on the keyboard) wouldn't show until the next input wakes the loop. This
-    /// schedules that follow-up frame so the change appears at once.
-    ///
-    /// Two passes, not one: this pass draws the output built before the command ran,
-    /// and the pass after it is the one the change is actually in. Marking only this
-    /// pass leaves that second one with nothing to make it dirty, and on a device
-    /// slow enough to have no spare passes the screen then trails the input.
+    /// Ask the loop to render one more frame without blocking on input: commands
+    /// drain *after* this frame's [`AppUi::update`], so a UI change they make
+    /// would otherwise wait for the next input. Two passes, not one — this pass
+    /// draws output built before the command ran; the next holds the change.
     #[inline]
     pub fn request_repaint(&mut self) {
         self.repaint_delay = Some(Duration::ZERO);
         self.forced_passes = 2;
     }
 
-    /// Whether the frame just built differs from the one already on the panel.
-    ///
-    /// Composing and presenting is the whole cost of a frame on a GPU-less
-    /// device — some 90 ms at 752x560, against under a millisecond for a page
-    /// that did not repaint — so a frame that would come out identical is worth
-    /// not drawing at all. egui reporting `MAX` means it is idle and would draw
-    /// the same picture again; an event it consumed, an animation it is running,
-    /// or the lingering cursor overlay all say otherwise.
+    /// Whether the frame just built differs from the one on the panel. Composing
+    /// and presenting is the whole frame cost on a GPU-less device (~90 ms at
+    /// 752x560), so an identical frame is worth not drawing. egui reporting `MAX`
+    /// means idle; a consumed event, an animation, or the cursor say otherwise.
     pub fn take_frame_dirty(&mut self, window: &AppWindow) -> bool {
         let forced = self.forced_passes > 0;
         self.forced_passes = self.forced_passes.saturating_sub(1);
@@ -411,210 +345,15 @@ impl AppUi {
             || self.cursor_visible_for().is_some()
     }
 
-    /// Move the gamepad cursor by a logical-px delta and mark it visible. Clamped
-    /// to the window (inset by the cursor's painted extent so the whole circle
-    /// stays on screen); it may roam over the toolbar so its buttons are clickable.
-    #[inline]
-    pub fn move_cursor(&mut self, dx: f32, dy: f32, window: &AppWindow) {
-        let (w, h) = window.size();
-        // The window in points: the cursor is drawn in them, the window is pixels.
-        let (w, h) = self.to_points(w as f32, h as f32);
-        self.cursor.0 = (self.cursor.0 + dx).clamp(CURSOR_EXTENT, w - CURSOR_EXTENT);
-        self.cursor.1 = (self.cursor.1 + dy).clamp(CURSOR_EXTENT, h - CURSOR_EXTENT);
-        self.cursor_last_move = Some(Instant::now());
-    }
-
-    /// Whether the cursor is over the web view (below the toolbar). Clicks there
-    /// go to the page; clicks above go to the egui toolbar via [`AppUi::click_ui`].
-    #[inline]
-    pub fn cursor_over_browser(&self) -> bool {
-        // Anywhere not on the toolbar chrome is the page (the web view fills the
-        // rest). When the toolbar is hidden, its rect is off-screen, so the whole
-        // window is the page.
-        !self
-            .toolbar_rect
-            .contains(egui::pos2(self.cursor.0, self.cursor.1))
-    }
-
-    /// Click the egui UI element under the cursor by feeding the backend a
-    /// synthetic mouse button event (egui never sees the gamepad otherwise).
-    /// `pressed` mirrors the A button's press/release so egui registers a click.
-    pub fn click_ui(&mut self, pressed: bool, window: &mut AppWindow) {
-        let ppp = self.egui_ctx.pixels_per_point();
-        let (x, y) = ((self.cursor.0 * ppp) as i32, (self.cursor.1 * ppp) as i32);
-        let window_id = window.sdl2_window().id();
-        let event = if pressed {
-            sdl2::event::Event::MouseButtonDown {
-                timestamp: 0,
-                window_id,
-                which: 0,
-                mouse_btn: sdl2::mouse::MouseButton::Left,
-                clicks: 1,
-                x,
-                y,
-            }
-        } else {
-            sdl2::event::Event::MouseButtonUp {
-                timestamp: 0,
-                window_id,
-                which: 0,
-                mouse_btn: sdl2::mouse::MouseButton::Left,
-                clicks: 1,
-                x,
-                y,
-            }
-        };
-        let _ = window.on_event(&event);
-        self.repaint_pending = true;
-    }
-
-    /// Time left before the cursor auto-hides, or `None` if it's already hidden
-    /// (never moved or idle past `cursor_linger`).
-    #[inline]
-    fn cursor_visible_for(&self) -> Option<Duration> {
-        self.cursor_last_move
-            .and_then(|t| self.cursor_linger.checked_sub(t.elapsed()))
-    }
-
-    /// The gamepad cursor in browser-relative coordinates (below the toolbar),
-    /// ready to feed to Servo as a mouse position.
-    #[inline]
-    pub fn cursor_browser_rel(&self) -> (f32, f32) {
-        self.to_browser_rel_pos(self.cursor.0, self.cursor.1)
-    }
-
-    /// The current input owner — see [`Focus`] for the precedence.
-    #[inline]
-    pub fn focus(&self) -> Focus {
-        if self.osk.visible {
-            Focus::Osk
-        } else if self.prompt.visible() {
-            Focus::Prompt
-        } else if self.menu.visible {
-            Focus::Menu
-        } else if self.settings.visible() {
-            Focus::Settings
-        } else if self.hints.visible {
-            Focus::Hints
-        } else if self.dial_edit.visible() {
-            Focus::DialEdit
-        } else if self.home_active {
-            Focus::Home
-        } else {
-            Focus::Page
-        }
-    }
-
-    /// Apply an [`OskCommand`] to the on-screen keyboard, routing typed input
-    /// to a modal `prompt()` dialog's field when one is up, else the address
-    /// bar when it holds focus, otherwise the focused page element.
-    pub fn osk(&mut self, cmd: OskCommand, browser: &AppBrowser, commands: &mut Vec<AppCommand>) {
-        let to_address_bar = self.address_bar_focused();
-        let target = if self.prompt.visible() && self.prompt.has_text_field() {
-            OskTarget::Prompt(self.prompt.input_mut())
-        } else if self.settings.visible() && self.settings.selected_is_text() {
-            // The settings overlay's focused text row: typing lands in the draft
-            // (the OSK only opens over a text row — see `App::settings_confirm`).
-            OskTarget::Settings(self.settings.selected_text_mut().expect("text row"))
-        } else if self.dial_edit.visible() {
-            // The speed-dial editor's URL field (its own buffer); Enter pins it.
-            OskTarget::DialEdit(self.dial_edit.input_mut())
-        } else if self.home_active {
-            // On the start page, typed text goes to its own search field, not
-            // the address bar (which only ever shows `retsurf:home` there).
-            OskTarget::Home(self.home.input_mut())
-        } else if to_address_bar {
-            OskTarget::AddressBar
-        } else {
-            OskTarget::Page
-        };
-        let to_page = matches!(target, OskTarget::Page);
-        self.osk.handle(cmd, target, browser, commands);
-        // A page field can sit under the keyboard, and only the page can scroll
-        // it out — the lift is applied once the keyboard's height is known (see
-        // `update`).
-        if to_page && matches!(cmd, OskCommand::Show) {
-            self.osk_lift_pending = true;
-        }
-    }
-
-    /// Open the menu. It takes over the stick and A, so the other user
-    /// overlays close — input focus and draw order can never disagree.
-    #[inline]
-    pub fn menu_open(&mut self) {
-        self.osk.visible = false;
-        self.hints.hide();
-        self.menu.open();
-    }
-
-    /// Open the settings overlay, seeding its draft from the live config. Like
-    /// the menu it takes over the stick and A, so the other user overlays close.
-    #[inline]
-    pub fn settings_open(&mut self, config: &AppConfig) {
-        self.osk.visible = false;
-        self.hints.hide();
-        self.menu.close();
-        self.settings.open(config);
-    }
-
-    /// Close the settings overlay, handing back its edited config and bindings
-    /// drafts so the app can save them and re-apply what changes live.
-    #[inline]
-    pub fn settings_close(&mut self) -> (AppConfig, Option<inputbind::Store>) {
-        let drafts = (self.settings.draft(), self.settings.changed_bindings());
-        self.settings.close();
-        drafts
-    }
-
-    /// Move the settings selection by `dy` rows. On the About tab the update block's
-    /// row count depends on the live update state (a release-notes link appears when
-    /// an update is available), so it's resolved here and handed to the nav.
-    #[inline]
-    pub fn settings_move(&mut self, dy: i32) {
-        let update_rows = settings::update_row_count(&self.update.snapshot());
-        self.settings.move_sel(dy, update_rows);
-    }
-
-    /// The action for the focused About-tab row on A: within the update block, the
-    /// primary action (row 0) or the "View release notes" link (row 1, when an update
-    /// is available); past it, the static link at `sel - update_rows`. `None` when the
-    /// row has nothing to do (a check / download / install in progress).
-    pub fn about_activate(&self) -> Option<SettingsAction> {
-        let sel = self.settings.selected();
-        let state = self.update.snapshot();
-        let update_rows = settings::update_row_count(&state);
-        if sel < update_rows {
-            if sel == 0 {
-                settings::update_command(&state)
-            } else {
-                settings::release_link(&state).map(SettingsAction::OpenLink)
-            }
-        } else {
-            crate::overlay::settings::about_info()
-                .links
-                .get(sel - update_rows)
-                .map(|(_, url)| SettingsAction::OpenLink(url.to_string()))
-        }
-    }
-
-    /// Set how long the gamepad cursor lingers after a move (the app calls this
-    /// when the interface config changes live via the settings overlay).
-    #[inline]
-    pub fn set_cursor_linger(&mut self, ms: u64) {
-        self.cursor_linger = Duration::from_millis(ms);
-    }
-
-    /// Move the toolbar to a window edge (the app calls this when the display
-    /// config changes live via the settings overlay). The next frame re-lays the
-    /// panel and the web view follows automatically.
+    /// Move the toolbar to a window edge (live config change). The next frame
+    /// re-lays the panel and the web view follows automatically.
     #[inline]
     pub fn set_toolbar_position(&mut self, pos: ToolbarPosition) {
         self.toolbar_position = pos;
     }
 
-    /// Enable/disable scroll-driven auto-hide (the app calls this on a live config
-    /// change). Re-showing the toolbar when turned off avoids leaving it stuck
-    /// hidden from a prior scroll.
+    /// Enable/disable scroll-driven auto-hide (live config change). Re-showing
+    /// the toolbar avoids leaving it stuck hidden from a prior scroll.
     #[inline]
     pub fn set_toolbar_autohide(&mut self, on: bool) {
         self.toolbar_autohide = on;
@@ -623,32 +362,9 @@ impl AppUi {
         }
     }
 
-    /// Whether hint mode draws combo badges (the app calls this on a live config
-    /// change). When off, hint mode is plain spatial hopping.
-    #[inline]
-    pub fn set_hint_badges(&mut self, on: bool) {
-        self.hint_badges = on;
-    }
-
-    /// Whether hint-mode combo badges are enabled (the router gates typed input
-    /// on the same flag).
-    #[inline]
-    pub fn hint_badges(&self) -> bool {
-        self.hint_badges
-    }
-
-    /// The event handler notes which device produced the latest input, so hint
-    /// mode can pick its badge alphabet (typed letters vs gamepad combos) when it
-    /// opens. `true` = keyboard.
-    #[inline]
-    pub fn note_input_keyboard(&mut self, keyboard: bool) {
-        self.last_input_keyboard = keyboard;
-    }
-
     /// Feed a page-scroll delta (the same `dy` handed to [`AppBrowser::scroll`]:
     /// positive reveals lower content) so the toolbar can hide on scroll-down and
-    /// reveal on scroll-up. Accumulates until a threshold to debounce jitter; a
-    /// no-op unless auto-hide is on.
+    /// reveal on scroll-up. Accumulates to a threshold; a no-op without auto-hide.
     pub fn notify_page_scroll(&mut self, dy: f32) {
         if !self.toolbar_autohide || dy == 0.0 {
             return;
@@ -669,218 +385,6 @@ impl AppUi {
         }
     }
 
-    /// Kick off a self-update check in the background (About tab; a no-op off a
-    /// PortMaster install). See [`crate::update`].
-    #[inline]
-    pub fn update_check(&self, sender: &UserEventSender) {
-        self.update.check(sender);
-    }
-
-    /// Startup: run a throttled background update check if `[update] auto_check` is
-    /// on and one is due (see [`crate::update::Updater::auto_check`]).
-    #[inline]
-    pub fn update_auto_check(&self, sender: &UserEventSender) {
-        self.update.auto_check(sender);
-    }
-
-    /// Download + install the available update in the background (About tab).
-    #[inline]
-    pub fn update_install(&self, sender: &UserEventSender) {
-        self.update.install(sender);
-    }
-
-    /// Adopt edited `[update]` settings live (settings overlay), for the next check.
-    #[inline]
-    pub fn set_update_config(&mut self, cfg: &UpdateConfig) {
-        self.update.set_config(cfg);
-    }
-
-    /// Mirror whether the active tab is on the start page (called each frame
-    /// from the main loop). Resets the overlay's state on entry so it always
-    /// opens focused on an empty search field. Returns whether the state
-    /// changed, so the caller can request a follow-up repaint: the start page
-    /// becomes active when an async navigation completes (not via a command or
-    /// input event), so the blocking idle loop would otherwise size the fresh
-    /// overlay invisibly and never paint its positioned frame.
-    #[inline]
-    pub fn set_home_active(&mut self, active: bool) -> bool {
-        let changed = active != self.home_active;
-        if active && !self.home_active {
-            self.home.reset();
-        }
-        self.home_active = active;
-        changed
-    }
-
-    /// Focus the start page's search field (when the OSK opens to type).
-    #[inline]
-    pub fn home_focus_search(&mut self) {
-        self.home.focus_search();
-    }
-
-    /// The start-page search field's current text (for submitting it from the
-    /// keyboard's Enter).
-    #[inline]
-    pub fn home_search_text(&self) -> String {
-        self.home.input().to_string()
-    }
-
-    /// Move the start-page selection by one dominant-axis step. The grid holds
-    /// one tile per pin plus a trailing "+ Add" tile, hence `len() + 1`.
-    #[inline]
-    pub fn home_move(&mut self, dx: i32, dy: i32) {
-        let count = self.menu.dial.urls().len() + 1;
-        self.home.move_sel(dx, dy, count);
-    }
-
-    /// The focused tile's pinned URL, if a *pin* tile is selected (the trailing
-    /// "Edit" tile has no URL — see [`Self::home_tile_is_edit`]).
-    #[inline]
-    pub fn home_selected_url(&self) -> Option<String> {
-        self.home
-            .tile()
-            .and_then(|i| self.menu.dial.urls().get(i).cloned())
-    }
-
-    /// Whether the trailing "Edit" tile (index == pin count) is focused.
-    #[inline]
-    pub fn home_tile_is_edit(&self) -> bool {
-        self.home.tile() == Some(self.menu.dial.urls().len())
-    }
-
-    // --- Speed-dial editor (the standalone overlay opened from the start page) ---
-
-    /// Open the speed-dial editor overlay.
-    #[inline]
-    pub fn open_pins_editor(&mut self) {
-        self.dial_edit.open();
-    }
-
-    /// Close the speed-dial editor (back to the start page).
-    #[inline]
-    pub fn close_pins_editor(&mut self) {
-        self.dial_edit.close();
-    }
-
-    /// The editor's focused item (drives the **A** action in the router).
-    #[inline]
-    pub fn dial_edit_item(&self) -> EditItem {
-        self.dial_edit.item()
-    }
-
-    /// Focus the editor's URL field (e.g. before opening the OSK to type).
-    #[inline]
-    pub fn dial_edit_focus_field(&mut self) {
-        self.dial_edit.focus_field();
-    }
-
-    /// The editor's URL field text (trimmed submission lives in the app).
-    #[inline]
-    pub fn dial_edit_input(&self) -> String {
-        self.dial_edit.input().to_string()
-    }
-
-    /// Clear the editor's URL field (after pinning its contents).
-    #[inline]
-    pub fn dial_edit_clear_input(&mut self) {
-        self.dial_edit.clear_input();
-    }
-
-    /// Move the editor's selection by one dominant-axis step.
-    #[inline]
-    pub fn dial_edit_move(&mut self, dx: i32, dy: i32) {
-        self.dial_edit.move_sel(dx, dy, self.dial_edit_slots());
-    }
-
-    /// The editor's focused pin index, if a tile (not the field) is focused.
-    #[inline]
-    pub fn dial_edit_tile(&self) -> Option<usize> {
-        self.dial_edit.tile()
-    }
-
-    fn dial_settings_pinned(&self) -> bool {
-        self.menu.dial.contains(crate::data::dial::SETTINGS_PIN)
-    }
-
-    #[inline]
-    fn dial_edit_slots(&self) -> usize {
-        dial_edit::slot_count(self.menu.dial.urls())
-    }
-
-    /// Whether the trailing "Pin settings" slot is focused — drives **A** in the
-    /// editor.
-    pub fn dial_edit_pin_settings_selected(&self) -> bool {
-        !self.dial_settings_pinned() && self.dial_edit.tile() == Some(self.menu.dial.urls().len())
-    }
-
-    /// Delete the editor's focused pin (X). The trailing tile's slot is out of
-    /// range, so it's a no-op there.
-    pub fn dial_edit_remove_selected(&mut self) {
-        if let Some(slot) = self.dial_edit.tile() {
-            self.menu.dial.remove(slot);
-        }
-    }
-
-    /// Move the editor's focused pin by `delta` slots (L1/R1), taking the
-    /// selection with it.
-    pub fn dial_edit_move_selected(&mut self, delta: i32) {
-        let Some(slot) = self.dial_edit.tile() else {
-            return;
-        };
-        let pins = self.menu.dial.urls().len();
-        let target = slot as i32 + delta;
-        if slot >= pins || target < 0 || target as usize >= pins {
-            return;
-        }
-        let target = target as usize;
-        self.menu.dial.swap(slot, target);
-        self.dial_edit.select_tile(target);
-    }
-
-    /// Whether a start-page tile (not the search field) is focused.
-    #[inline]
-    pub fn home_tile_selected(&self) -> bool {
-        self.home.tile().is_some()
-    }
-
-    /// Hint mode opened: a collection round was started in the browser. The badge
-    /// alphabet follows the device that triggered it (typed letters from the
-    /// keyboard, button combos from the pad) — fixed for the whole session.
-    #[inline]
-    pub fn hints_begin_collect(&mut self) {
-        let labels = if self.last_input_keyboard {
-            HintLabels::Keyboard
-        } else {
-            HintLabels::Gamepad
-        };
-        self.hints.begin_collect(labels);
-    }
-
-    /// Fresh clickable rects from the page. Selection lands near the previous
-    /// one (a post-scroll refresh) or near the gamepad cursor (mode entry).
-    pub fn hints_apply(&mut self, rects: Vec<Hint>) {
-        let near = self
-            .hints
-            .selected_center()
-            .unwrap_or_else(|| self.cursor_browser_rel());
-        // Combo codes are ordered from the viewport top-center (browser-relative),
-        // so the nearest-to-top hints get the shortest codes.
-        let top_center = (self.webview_rect.width() / 2.0, 0.0);
-        self.hints.show(rects, near, top_center);
-    }
-
-    /// Feed one gamepad combo symbol into hint mode (see [`HintInput`]).
-    #[inline]
-    pub fn hints_push_sym(&mut self, s: Sym) -> HintInput {
-        self.hints.push_label(Label::Sym(s))
-    }
-
-    /// Feed one typed keyboard letter into hint mode (see [`HintInput`]).
-    #[inline]
-    pub fn hints_push_key(&mut self, c: char) -> HintInput {
-        self.hints.push_label(Label::Key(c))
-    }
-
     /// Height of the browser viewport (logical px) — for screen-relative scrolls.
     #[inline]
     pub fn browser_area_height(&self) -> f32 {
@@ -892,69 +396,6 @@ impl AppUi {
     #[inline]
     pub fn browser_area_width(&self) -> f32 {
         self.webview_rect.width()
-    }
-
-    /// Mirror the gamepad's latched D-pad scroll mode (called by the router
-    /// every analog frame) so the indicator tracks it. Entering the mode pings
-    /// the linger timer so the indicator shows up like the cursor would, then
-    /// auto-hides unless scrolling keeps it alive (see [`Self::mark_cursor_active`]).
-    #[inline]
-    pub fn set_scroll_mode(&mut self, on: bool) {
-        if on && !self.scroll_mode {
-            self.cursor_last_move = Some(Instant::now());
-        }
-        self.scroll_mode = on;
-    }
-
-    /// Refresh the linger timer without moving the cursor — used by active page
-    /// scroll so the scroll-mode indicator stays visible while scrolling, then
-    /// lingers and auto-hides exactly like the cursor.
-    #[inline]
-    pub fn mark_cursor_active(&mut self) {
-        self.cursor_last_move = Some(Instant::now());
-    }
-
-    /// Whether the address-bar text field currently holds keyboard focus (also
-    /// guards plain-key keyboard shortcuts in the event handler).
-    pub fn address_bar_focused(&self) -> bool {
-        self.egui_ctx
-            .memory(|m| m.has_focus(egui::Id::new("location")))
-    }
-
-    /// Which egui text field the OSK is currently typing into — mirrors the
-    /// target priority in [`AppUi::osk`], collapsing the no-egui-caret cases
-    /// (Page, settings rows) to `None`. Used to park that field's caret at the
-    /// buffer end while the OSK is up (see [`OskField`]).
-    fn osk_target_field(&self) -> OskField {
-        if !self.osk.visible {
-            OskField::None
-        } else if self.prompt.visible() && self.prompt.has_text_field() {
-            OskField::Prompt
-        } else if self.dial_edit.visible() {
-            OskField::DialEdit
-        } else if self.home_active {
-            OskField::Home
-        } else if self.address_bar_focused() {
-            OskField::AddressBar
-        } else {
-            OskField::None
-        }
-    }
-
-    /// Whether the start page's search field holds egui keyboard focus (a desktop
-    /// click into it). While it does, arrow keys edit text rather than moving the
-    /// start-page selection, and plain-key shortcuts are muted.
-    pub fn home_field_editing(&self) -> bool {
-        self.egui_ctx
-            .memory(|m| m.has_focus(egui::Id::new("home_search")))
-    }
-
-    /// Whether the speed-dial editor's URL field holds egui keyboard focus —
-    /// like [`Self::home_field_editing`], but for the editor's `dial_edit_url`
-    /// field.
-    pub fn dial_edit_field_editing(&self) -> bool {
-        self.egui_ctx
-            .memory(|m| m.has_focus(egui::Id::new("dial_edit_url")))
     }
 
     /// A window pixel (as SDL reports events) in web-view points — the space the
@@ -975,38 +416,10 @@ impl AppUi {
         (dx / ppp, dy / ppp)
     }
 
-    /// Scale the whole chrome to the panel it is on, then by what the setting
-    /// asks for. Every size in the renderers is in points against a 640x480
-    /// design, so one zoom factor carries the lot. The page follows on the same
-    /// factor, so a CSS pixel and a point stay the same size.
-    fn sync_scale(&mut self, browser: &AppBrowser) {
-        let ctx = &self.egui_ctx;
-        let installed = ctx.zoom_factor();
-        let native = panel_points(ctx.content_rect().size(), installed);
-        if !(1.0..=MAX_PANEL).contains(&native.x) || !(1.0..=MAX_PANEL).contains(&native.y) {
-            return;
-        }
-        let wanted = wanted_scale(native, self.forced_scale, self.ui_scale);
-        browser.set_hidpi(wanted);
-        if (wanted - installed).abs() < f32::EPSILON {
-            return;
-        }
-        // Two decimals: snapping to the step grid leaves float noise in the tail.
-        log::info!("ui scale {wanted:.2} for {}x{} points", native.x, native.y);
-        ctx.set_zoom_factor(wanted);
-        self.forced_passes = self.forced_passes.max(1);
-    }
-
-    /// Adopt an edited `[display] scale`; the next frame installs it.
-    pub fn set_ui_scale(&mut self, scale: f32) {
-        self.ui_scale = scale;
-    }
-
-    /// Resize the browser to the central web-view area (the window minus the
-    /// toolbar) for the current window size. Driven by SDL window-resize events:
-    /// egui's reactive sizing reads the central rect a frame later, so it can lag
-    /// behind the actual window. Uses the toolbar height measured during `update`,
-    /// and shares `browser_viewport` with that path so the two never double-resize.
+    /// Resize the browser to the web-view area on SDL window-resize events:
+    /// egui's reactive sizing reads the central rect a frame later and lags the
+    /// window. Uses the toolbar height measured in [`AppUi::update`] and shares
+    /// `browser_viewport` with it, so the two never double-resize.
     pub fn resize_browser(&mut self, window: &AppWindow, browser: &AppBrowser) {
         let (dw, dh) = window.drawable_size();
         if dw == 0 || dh == 0 {
@@ -1028,11 +441,9 @@ impl AppUi {
         }
     }
 
-    /// Refresh egui's cached window size from the live window. Called once per
-    /// frame so an orientation change is reflected even when SDL doesn't deliver
-    /// a size-changed event on Android rotation (otherwise the UI keeps laying
-    /// out for the previous orientation — e.g. a landscape home page shown in
-    /// portrait).
+    /// Refresh egui's cached window size from the live window, once per frame:
+    /// SDL doesn't deliver a size-changed event on Android rotation, and the UI
+    /// would keep laying out for the previous orientation.
     #[cfg(target_os = "android")]
     pub fn sync_window_size(&mut self, window: &mut AppWindow) {
         window.sync_egui_window_size();
@@ -1047,10 +458,8 @@ impl AppUi {
     }
 
     /// Fold the idle-repaint sources into `repaint_delay` so the blocking wait
-    /// wakes on its own when something time-based comes due: the lingering cursor
-    /// needs erasing (`cursor_visible`, the soonest of the linger end), a
-    /// post-scroll hint refresh is pending, or the debug memory overlay wants its
-    /// ~1 Hz tick. Each only ever shortens the wait; `None` leaves it untouched.
+    /// wakes on its own when something time-based comes due. Each source only
+    /// ever shortens the wait; `None` leaves it untouched.
     fn schedule_idle_repaints(&mut self, cursor_visible: Option<Duration>) {
         self.repaint_delay = cursor_visible;
         // A pending post-scroll hint refresh also needs the loop to wake by
@@ -1066,12 +475,10 @@ impl AppUi {
         }
     }
 
-    /// Snapshot the per-frame browser/overlay inputs the render closure needs, as
-    /// owned copies — so the browser borrow doesn't leak into `egui.run` (which
-    /// borrows `self`). The tab reads happen here, *before* the long
-    /// `get_state_mut` borrow taken just before `run`: both touch the browser's
-    /// tab list and can't overlap. `osk_field`/`osk_caret` are pure `self` reads,
-    /// gathered here too since the closure can't take all of `self`.
+    /// Snapshot the per-frame browser/overlay inputs as owned copies, so the
+    /// browser borrow doesn't leak into `egui.run` (which borrows `self`). The
+    /// tab reads happen *before* the long `get_state_mut` borrow taken just
+    /// before `run`: both touch the browser's tab list and can't overlap.
     fn frame_snapshot(&mut self, browser: &mut AppBrowser) -> FrameInputs {
         let tab_count = browser.tab_count();
         let zoom_pct = browser.zoom_chip();
@@ -1091,9 +498,8 @@ impl AppUi {
     }
 
     /// Keep the start-page / speed-dial-editor selections in range before they
-    /// render. The pin list itself isn't snapshotted — the overlays borrow it
-    /// straight from the live store at their call sites, so there's no per-frame
-    /// clone of the (kept-in-sync) speed-dial Vec.
+    /// render. The pin list isn't snapshotted — the overlays borrow it straight
+    /// from the live store at their call sites.
     fn clamp_overlay_selections(&mut self) {
         let pin_count = self.menu.dial.urls().len();
         if self.home_active {
@@ -1108,11 +514,11 @@ impl AppUi {
         }
     }
 
-    /// Decide the toolbar layout for this frame, before the egui closure (these
-    /// reads — esp. `focus()` — borrow all of `self`, which can't overlap
-    /// `egui.run`). Auto-hide forces the bar visible while typing, and floats it
-    /// on either edge: a strip that came and went would resize the web view, and
-    /// that is a full Servo reflow mid-scroll. Otherwise the bar is a panel.
+    /// Decide the toolbar layout before the egui closure (these reads — esp.
+    /// `focus()` — borrow all of `self`, which can't overlap `egui.run`).
+    /// Auto-hide forces the bar visible while typing and floats it on either
+    /// edge: a strip that came and went would resize the web view, and that is
+    /// a full Servo reflow mid-scroll. Otherwise the bar is a panel.
     fn toolbar_layout(&self) -> ToolbarLayout {
         let typing = self.focus() != Focus::Page;
         ToolbarLayout {
@@ -1137,9 +543,8 @@ impl AppUi {
         // Before the layout: every rect below is measured in the zoom in force.
         self.sync_scale(browser);
 
-        // The cursor draws only while it lingers after a move. When it does, ask
-        // the loop to wake when the linger ends so it gets erased even if no other
-        // event arrives; otherwise leave the idle wait untouched.
+        // The cursor draws only while it lingers after a move; ask the loop to
+        // wake when the linger ends so it gets erased without another event.
         let cursor_visible = if self.focus() == Focus::Page {
             self.cursor_visible_for()
         } else {
@@ -1152,10 +557,8 @@ impl AppUi {
 
         {
             let mut state = browser.get_state_mut();
-            // Owned copies (see `FrameInputs`), so the browser borrow above and the
-            // `self` reads here don't leak into the closure below. `caret_for`
-            // parks each `TextEdit`'s cursor at the OSK caret for the field it
-            // types into; the toolbar layout was decided in `toolbar_layout`.
+            // Owned copies (see `FrameInputs`); `caret_for` parks each
+            // `TextEdit`'s cursor at the OSK caret for the field it types into.
             let FrameInputs {
                 tab_count,
                 zoom_pct,
@@ -1172,8 +575,7 @@ impl AppUi {
             // Snapshot the self-update state here (the About tab reads it): the
             // `self.update` borrow can't overlap the `self`-borrowing closure below.
             let update = self.update.snapshot();
-            // Toolbar update chip: shown only once a background/manual check has
-            // found a newer build.
+            // Toolbar update chip: shown only once a check has found a newer build.
             let update_available = matches!(update, UpdateState::Available { .. });
             window.run_ui(|ctx| {
                 let ppp = ctx.pixels_per_point();
@@ -1187,9 +589,8 @@ impl AppUi {
                 let bookmarked = self.menu.is_bookmarked(state.get_location());
                 let active_downloads = self.menu.downloads.active_count();
 
-                // 1) Reserved-space toolbar: the panel reserves its strip and the
-                //    page reflows below it. Only when auto-hide is off — an
-                //    auto-hiding bar floats instead, so it never resizes the page.
+                // 1) Reserved-space toolbar (auto-hide off): the panel reserves
+                //    its strip and the page reflows below it.
                 if !toolbar_overlay && toolbar_shown {
                     self.toolbar_rect = toolbar::add_toolbar(
                         &mut root,
@@ -1242,8 +643,7 @@ impl AppUi {
                 // 3) Floating overlay toolbar (auto-hide, either edge): over the
                 //    full-height web view, so toggling it resizes nothing.
                 if toolbar_overlay {
-                    // Instant show/hide: draw the bar only while shown, and skip it
-                    // entirely while hidden — a foreground `Area` costs a
+                    // Draw only while shown: a foreground `Area` costs a
                     // tessellation pass every frame even off-screen.
                     if toolbar_shown {
                         self.toolbar_rect = toolbar::add_toolbar_overlay(
@@ -1265,11 +665,9 @@ impl AppUi {
                     }
                 }
 
-                // The start-page overlay is a backdrop over the (blank) web
-                // view — drawn below the foreground overlays (menu / OSK / etc.)
-                // so they can still open on top of it. The dial editor (also
-                // reached from the start page) fully covers it, so skip the start
-                // page underneath while the editor is up.
+                // The start page is a backdrop over the (blank) web view, drawn
+                // below the foreground overlays so they can open on top. The dial
+                // editor fully covers it, so skip it underneath.
                 if self.home_active && !self.dial_edit.visible() {
                     home::add_home(
                         ctx,
@@ -1293,15 +691,11 @@ impl AppUi {
                     );
                 }
 
-                // The settings overlay: a full-screen panel like the menu. Drawn
-                // before the menu/OSK chain below so the OSK can open on top to
-                // type into a text field (focus is `Osk` then, painting it here
-                // would put it under the keyboard).
+                // Settings: drawn before the menu/OSK chain so the OSK can open
+                // on top to type into a text field.
                 if self.settings.visible() {
-                    // Neither overlay has a text field, and both own their own
-                    // row selection — an egui-focused row (Tab, our section
-                    // switch, moves the focus) would take Enter a second time
-                    // and activate twice.
+                    // The overlay owns its row selection; an egui-focused row
+                    // would take Enter a second time and activate twice.
                     drop_egui_focus(ctx);
                     settings::add_settings(ctx, &self.settings, &update, commands);
                 }
@@ -1338,30 +732,8 @@ impl AppUi {
                 } else if self.hints.visible {
                     hints::add_hints(ctx, &self.hints, self.webview_rect, self.hint_badges);
                 } else if cursor_visible.is_some() {
-                    // Gamepad cursor overlay, always on top. `cursor` is in logical
-                    // px which equals egui points at the handheld's 1.0 scale factor.
-                    let painter = ctx.layer_painter(egui::LayerId::new(
-                        egui::Order::Foreground,
-                        egui::Id::new("gamepad_cursor"),
-                    ));
                     let pos = egui::pos2(self.cursor.0, self.cursor.1);
-                    if self.scroll_mode {
-                        // Autoscroll-style indicator (dot + up/down arrowheads).
-                        // Follows the same linger/auto-hide as the cursor: shown on
-                        // entering scroll mode and while scrolling, then fades out.
-                        add_scroll_indicator(&painter, pos);
-                    } else {
-                        painter.circle_filled(
-                            pos,
-                            CURSOR_RADIUS,
-                            egui::Color32::from_white_alpha(235),
-                        );
-                        painter.circle_stroke(
-                            pos,
-                            CURSOR_RADIUS,
-                            egui::Stroke::new(CURSOR_STROKE, egui::Color32::BLACK),
-                        );
-                    }
+                    cursor::paint_cursor(ctx, pos, self.scroll_mode);
                 }
 
                 // Debug memory overlay (opt-in), drawn last so it sits above
@@ -1389,10 +761,9 @@ impl AppUi {
             self.osk_lift_pending = false;
         }
 
-        // Fold in egui's own repaint timing. A freshly shown anchored `Area`
-        // (the menu / OSK) sizes itself invisibly on its first frame and asks
-        // egui for an immediate follow-up to paint it positioned — without this
-        // the loop blocks on input and the overlay only appears after the next
+        // Fold in egui's own repaint timing: a freshly shown anchored `Area`
+        // sizes itself invisibly and asks for an immediate follow-up to paint
+        // positioned — without this the overlay appears only after the next
         // keypress. `MAX` means egui is idle, so it never shortens our wait.
         let egui_delay = window.repaint_delay();
         if egui_delay < Duration::MAX {
@@ -1426,173 +797,12 @@ impl AppUi {
         self.toolbar_rect.contains(pos)
     }
 
-    /// Whether a *pixel*-space y coordinate (as carried by raw SDL finger events)
-    /// lands in the web-view area, below the toolbar. Touches over the toolbar are
-    /// egui's — it synthesizes pointer events from them — so only web-view touches
-    /// should start a page scroll/tap gesture. The web-view rect is in egui
-    /// points, so scale it up to compare against the pixel coordinate.
+    /// Whether a *pixel*-space y (raw SDL finger events) lands in the web view,
+    /// below the toolbar: touches over the toolbar are egui's, so only web-view
+    /// touches should start a page scroll/tap gesture.
     #[inline]
     pub fn point_over_webview(&self, y_px: f32) -> bool {
         let y = y_px / self.egui_ctx.pixels_per_point();
         !self.toolbar_rect.y_range().contains(y)
-    }
-}
-
-/// The panel behind a laid-out rect, in points at zoom 1: the drawn area with the
-/// zoom divided back out, rounded so the measurement does not feed on its own
-/// output and swap between two steps every frame.
-fn panel_points(content: egui::Vec2, zoom: f32) -> egui::Vec2 {
-    (content * zoom).round()
-}
-
-/// The zoom a panel of `native` points wants: what the design fits into it, or
-/// `forced` where a launcher said so, taken `user` times over and snapped to a
-/// step the layout can settle on.
-fn wanted_scale(native: egui::Vec2, forced: Option<f32>, user: f32) -> f32 {
-    // Long edge to long edge: a turned screen is the same screen, so the chrome
-    // keeps its size and the page gets the spare measure instead.
-    let fit =
-        (native.max_elem() / BASE_SIZE.max_elem()).min(native.min_elem() / BASE_SIZE.min_elem());
-    let fit = if fit >= 1.0 && fit.fract() <= WHOLE_ZOOM_REACH {
-        fit.floor()
-    } else {
-        fit
-    };
-    let wanted = forced.unwrap_or(fit) * user;
-    ((wanted / SCALE_STEP).round() * SCALE_STEP).clamp(*SCALE_RANGE.start(), *SCALE_RANGE.end())
-}
-
-/// The D-pad scroll-mode indicator at the parked cursor position: a center dot
-/// with up/down arrowheads, like a browser's middle-click autoscroll marker.
-fn add_scroll_indicator(painter: &egui::Painter, pos: egui::Pos2) {
-    let fill = egui::Color32::from_white_alpha(235);
-    let stroke = egui::Stroke::new(CURSOR_STROKE, egui::Color32::BLACK);
-    painter.circle_filled(pos, 2.5, fill);
-    painter.circle_stroke(pos, 2.5, stroke);
-    for dir in [-1.0f32, 1.0] {
-        let tip = egui::pos2(pos.x, pos.y + dir * 12.0);
-        let base = pos.y + dir * 5.5;
-        let points = vec![
-            tip,
-            egui::pos2(pos.x - 4.5, base),
-            egui::pos2(pos.x + 4.5, base),
-        ];
-        painter.add(egui::Shape::convex_polygon(points, fill, stroke));
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::config::bounds::SCALE;
-
-    /// The panel the chrome was drawn for, the Miyoo Flip's, and a desktop window.
-    const HANDHELD: egui::Vec2 = BASE_SIZE;
-    const FLIP: egui::Vec2 = egui::vec2(752.0, 560.0);
-    const DESKTOP: egui::Vec2 = egui::vec2(1280.0, 720.0);
-    /// The same handheld held the other way up.
-    const TURNED: egui::Vec2 = egui::vec2(BASE_SIZE.y, BASE_SIZE.x);
-
-    /// Snapping multiplies the step back out, so the answers land a rounding
-    /// short of the round number they read as.
-    fn assert_scale(got: f32, want: f32) {
-        assert!((got - want).abs() < SCALE_STEP / 2.0, "{got} is not {want}");
-    }
-
-    /// Every scale the setting can be stepped to.
-    fn user_scales() -> impl Iterator<Item = f32> {
-        let (min, max, step) = (SCALE.min as f32, SCALE.max as f32, SCALE_STEP);
-        let steps = ((max - min) / step).round() as i32;
-        (0..=steps).map(move |i| min + i as f32 * step)
-    }
-
-    #[test]
-    fn the_design_lands_at_one_on_the_panel_it_was_drawn_for() {
-        assert_scale(wanted_scale(HANDHELD, None, 1.0), 1.0);
-        assert_scale(wanted_scale(DESKTOP, None, 1.0), 1.5);
-    }
-
-    /// The Flip's panel is 17% past the design, which is not worth rendering type
-    /// between pixels for; the spare pixels widen the page instead.
-    #[test]
-    fn a_panel_a_little_past_the_design_keeps_a_whole_zoom() {
-        assert_scale(wanted_scale(FLIP, None, 1.0), 1.0);
-    }
-
-    #[test]
-    fn a_quarter_turn_hands_back_the_same_screen_and_the_same_sized_chrome() {
-        assert_scale(wanted_scale(TURNED, None, 1.0), 1.0);
-        for panel in [HANDHELD, FLIP, DESKTOP] {
-            let turned = egui::vec2(panel.y, panel.x);
-            for user in user_scales() {
-                assert_eq!(
-                    wanted_scale(panel, None, user),
-                    wanted_scale(turned, None, user),
-                    "{panel:?} at {user} resizes itself when turned"
-                );
-            }
-        }
-    }
-
-    #[test]
-    fn the_setting_is_read_against_whatever_the_panel_asked_for() {
-        assert_scale(wanted_scale(HANDHELD, None, 1.2), 1.2);
-        assert_scale(wanted_scale(DESKTOP, None, 1.2), 1.8);
-        // A launcher's pin (Android's density) is a base like any other.
-        assert_scale(wanted_scale(DESKTOP, Some(2.0), 1.2), 2.4);
-    }
-
-    #[test]
-    fn neither_end_of_the_setting_can_push_the_zoom_out_of_range() {
-        for panel in [HANDHELD, DESKTOP, egui::vec2(7680.0, 4320.0)] {
-            for user in [SCALE.min as f32, SCALE.max as f32] {
-                let scale = wanted_scale(panel, None, user);
-                assert!(
-                    SCALE_RANGE.contains(&scale),
-                    "{panel:?} at {user} -> {scale}"
-                );
-            }
-        }
-    }
-
-    /// What `Context::content_rect` hands back for a panel at `zoom`: the points
-    /// it lays out in, snapped to egui's 1/32pt grid.
-    fn content_rect(panel: egui::Vec2, zoom: f32) -> egui::Vec2 {
-        use egui::emath::GuiRounding as _;
-        (panel / zoom).round_ui()
-    }
-
-    /// One frame: measure the panel through the zoom in force, ask for the next.
-    fn next_zoom(panel: egui::Vec2, zoom: f32, user: f32) -> f32 {
-        wanted_scale(panel_points(content_rect(panel, zoom), zoom), None, user)
-    }
-
-    #[test]
-    fn a_scale_settles_rather_than_swapping_between_two_steps_every_frame() {
-        for panel in [HANDHELD, FLIP, TURNED, DESKTOP] {
-            for user in user_scales() {
-                let mut zoom = 1.0;
-                for _ in 0..8 {
-                    zoom = next_zoom(panel, zoom, user);
-                }
-                assert_eq!(
-                    next_zoom(panel, zoom, user),
-                    zoom,
-                    "{panel:?} at {user} swaps"
-                );
-            }
-        }
-    }
-
-    /// A setting stepping by 5% into a zoom snapped to 10% spends half its presses
-    /// on nothing at all.
-    #[test]
-    fn every_step_of_the_setting_moves_the_panel_the_design_was_drawn_for() {
-        let mut zooms: Vec<f32> = user_scales()
-            .map(|user| wanted_scale(HANDHELD, None, user))
-            .collect();
-        let asked = zooms.len();
-        zooms.dedup();
-        assert_eq!(zooms.len(), asked, "{zooms:?} repeats a zoom");
     }
 }
